requirements.txt
fastapi
uvicorn
load_dotenv
requests
line-bot-sdk
